match_case_calculator.py_2

+++ control-flow/match_case_calculator.py
@@ -1,5 +1,5 @@
-first_number = int(input("Enter the first number: "))
-second_number = int(input("Enter the second number: "))
+first_number = float(input("Enter the first number: "))
+second_number = float(input("Enter the second number: "))
 operation = input("Choose the operation (+, -, *, /): ")
 
 match operation:
